# docs/swagger.yaml
definitions:
  dtos.CartItemCartFormatter:
    properties:
      id:
        type: integer
      user_id:
        type: integer
    type: object
  dtos.CartItemFormatter:
    properties:
      cart:
        $ref: '#/definitions/dtos.CartItemCartFormatter'
      cart_id:
        type: integer
      id:
        type: integer
      price:
        type: number
      product:
        $ref: '#/definitions/dtos.CartProductFormatter'
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: number
      user_id:
        type: integer
    type: object
  dtos.CartProductCategoryFormatter:
    properties:
      category_id:
        type: integer
      description:
        type: string
      product_name:
        type: string
      user:
        $ref: '#/definitions/dtos.CartProductCategoryUserFormatter'
    type: object
  dtos.CartProductCategoryUserFormatter:
    properties:
      full_name:
        type: string
      role:
        type: string
      user_id:
        type: integer
    type: object
  dtos.CartProductFormatter:
    properties:
      category:
        $ref: '#/definitions/dtos.CartProductCategoryFormatter'
      description:
        type: string
      price:
        type: number
      product_id:
        type: integer
      product_name:
        type: string
    type: object
...
  dtos.CreateCartItemInput:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  utils.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
info:
  contact:
    email: nadapdaponai21@gmail.com
    name: API Support
    url: https://onainadapdap1.github.io/
  description: This is service for my Online Store API assignment.
  termsOfService: http://swagger.io/terms/
  title: Online Store API
  version: "1.0"
paths:
  /api/v1/carts:
    get:
      description: Get All user item cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CartItemFormatter'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get All user item cart
      tags:
      - carts
  /api/v1/carts/cart:
    post:
      consumes:
      - application/json
      description: Add product to cart
      parameters:
      - description: add product to cart input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCartItemInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CartItemFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Add product to cart
      tags:
      - carts
  /api/v1/carts/cart/{cart_id}/productID/{item_id}:
    delete:
      description: Delete item from cart
      parameters:
      - description: cart id
        in: path
        name: cart_id
        required: true
        type: integer
      - description: item id
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete item from cart
      tags:
      - carts
    put:
      consumes:
      - application/json
      description: Update cart item quantity based on action (add/remove) and quantity
      parameters:
      - description: Cart ID
        in: path
        name: cart_id
        required: true
        type: integer
      - description: Cart Item ID
        in: path
        name: item_id
        required: true
        type: integer
      - description: Action to perform (add/remove)
        in: formData
        name: action
        required: true
        type: string
      - description: Quantity to add/remove
        in: formData
        name: quantity
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CartItemFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - carts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"